requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
tensorflow==2.16.1  # Or pytorch
requests==2.31.0
influxdb-client==1.41.0
SQLAlchemy==2.0.29
matplotlib==3.8.4
seaborn==0.13.2
scipy==1.13.0
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.1
psycopg2-binary
pyyaml
tqdm
loguru
logzero==1.7.0
smartapi-python==1.5.5
pyotp==2.9.0
websocket-client==1.9.0
playwright==1.44.0
colorama
kiteconnect==5.0.1
